+++ internal/storage/pg/pg.go
@@ -112,7 +112,7 @@ func (p *PGDB) UpdateStatus(ctx context.Context, newStatus, order, user string) 
 	return nil
 }
 
-func (p *PGDB) UpdateUserBalance(ctx context.Context, user string, accrual, withdrawn float32) error {
+func (p *PGDB) UpdateUserBalance(ctx context.Context, user string, accrual, withdrawn float64) error {
 	query := `UPDATE users SET accrual = $1, withdrawn = $2
 			WHERE username = $3`
 	result, err := p.db.Exec(ctx, query, int(accrual*100), int(withdrawn*1000), user)
@@ -172,7 +172,7 @@ func (p *PGDB) GetUserBalance(ctx context.Context, user string) (models.UserBala
 	return balance, nil
 }
 
-func (p *PGDB) SetUserWithdrawn(ctx context.Context, orderNum, user string, withdrawn float32) error {
+func (p *PGDB) SetUserWithdrawn(ctx context.Context, orderNum, user string, withdrawn float64) error {
 	query := `INSERT INTO withdrawals (orderNum, sum, precessed_at, username)
 				VALUES ($1, $2, $3, $4) ON CONFLICT (orderNum) DO NOTHING`
 	_, err := p.db.Exec(ctx, query, orderNum, withdrawn*100, time.Now(), user)
